refactor: rename current theme to legacy, adapt imports, linter
